fix: Performance test failure when vela backend installed
Running test_backend_vela_performance.py directly via "pytest"
with the vela backend installed results in two failures. If the
backend is not installed there are no failures. If the tests are
run via "tox" then there are no failures.

The intent of the failing test is: given layer performance data
in CSV format, ensure that we can parse this to create a valid
LayerwisePerfInfo object. However, the parsing code appears to
have updated to cope with different versions of the underlying
vela backend tool; versions that produce data with different
field names (column headings). The test is failing because it
checks the __repr__ for the layer performance data classes,
which now contain a list of field name aliases
["TFLite_operator", "Original Operator"] instead of a single
name.

After discussion with Wojciech our conclusion was that the test
should use direct attribute checks. It should specify both the
list of field names and the corresponding list of actual fields
(attributes) to validate. The code under test, performance.py,
should also be updated as the __repr__ functions for the
layerwise performance data classes are non-standard and only
used in the tests. They should be removed entirely.

Note that a temporary fix for these failures was submitted as
part of MLIA-1423.

Resolves: MLIA-1552
Change-Id: I3d4cb9ea7bab71857690d6abf9105d5a23d3fade
Signed-off-by: James Edgar <james.edgar@arm.com>
Reviewed-on: https://eu-gerrit-2.euhpc.arm.com/c/ml/ecosystem/mlia/+/1149746
Tested-by: expkit <svc_expkit@arm.com>
Reviewed-by: Wojciech Boncela <wojciech.boncela@arm.com>
IP-review: Isabella Gottardi <isabella.gottardi@arm.com>
Reviewed-on: https://eu-gerrit-2.euhpc.arm.com/c/ml/ecosystem/mlia/+/1151451
Reviewed-by: Isabella Gottardi <isabella.gottardi@arm.com>